WKLoginView code added
